+++ src/ApiClientCodeGen.VSMac/CustomTools/OpenApiDocumentFactory.cs
@@ -1,7 +1,7 @@
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core.Generators.NSwag;
 using NSwag;
 
-namespace ApiClientCodeGen.VSMac.CustomTools.NSwag
+namespace ApiClientCodeGen.VSMac.CustomTools
 {
     public class OpenApiDocumentFactory : IOpenApiDocumentFactory
     {
